Improve code quality (#1460)
* ci: add .deepsource.toml

* refactor: remove unnecessary use of comprehension

It is unnecessary to use a comprehension just to loop over the `iterable` and create a `list`/`set`/`dict` out of it. Python has a specialized set of tools for this task: the `list`/`set`/`dict` constructors, which are faster and more readable.

* refactor: use `sys.exit()` calls

The `exit` or `quit` functions don't exist at top-level if python is started with the `-S` flag, and will raise an error. Use `sys.exit()` instead.

* refactor: use shallow copy of dictionary for iteration

Dictionaries are represented by a hash table and adding or removing items while iterating over it will alter the iteration order.
This will cause a `RuntimeError`.

* refactor: fix dangerous default argument

Do not use a mutable like `list` or `dictionary` as a default value to an argument. Python’s default arguments are evaluated once when the function is defined. Using a mutable default argument and mutating it will mutate that object for all future calls to the function as well.

* refactor: remove assert statement from non-test files

Usage of `assert` statement in application logic is discouraged. `assert` is removed with compiling to optimized byte code. Consider raising an exception instead. Ideally, `assert` statement should be used only in tests.

* u

* u

* refactor: remove unnecessary f-string

It is wasteful to use `f-string` mechanism if there are no expressions to be extrapolated. It is recommended to use regular strings instead.

* u

* u

* refactor: remove implicit `object` from the base class

The class is inheriting from `object`, which is implicit under Python 3 , hence can be safely removed from bases.  Not preferred:

* refactor: refactor `if` expression

The `if` expression used here can be simplified to increase the code readability. Example:

* refactor: merge `isinstance` calls

You can pass a tuple of types you want to check as the second argument to `isinstance`. If the object matches with any of the types, it will return `True` else `False`.

* refactor: remove unnecessary generator

It is unnecessary to use `list`, `set`, `dict` around a generator expression to get an object of that type since there are comprehensions for these types.

* refactor: use `max` built-in to get the maximum of two values

It is unnecessary to use an `if` statement to check the maximum of two values and then assign the value to a name. You can use the `max` built-in do do this. It is straightforward and more readable.

* refactor: simplify the `if` statement

The highlighted `if` block can be refactored into a simplified statement, increasing its readability.

* Update __init__.py

* refactor: remove redundant `None` default

It is unnecessary to provide `None` as the default value when the key is not present in the dictionary as `get` implicitly returns `None`.  Not preferred:

* refactor: iterate dictionary directly

Consider iterating the dictionary directly instead of calling `.keys()`.
Using `for key in dictionary` would always iterate the dictionary keys.

* u

* u

* u

* refactor: remove assignment of variable to itself

Variable is being assigned to itself.

* refactor: refactor for loop into `any` call

It is recommended to use `any` for this loop, as it will make your code smaller and more readable.

* refactor: replace range(len(...)) with enumerate(...)

Using `range(len(...))` is not pythonic. Python does not have not index-based loops. Instead, it uses collection iterators.  Python has a built-in method `enumerate` which adds a counter to an iterable.

* Update chunithm.py

* Update __init__.py

* Update __init__.py

* Update maimai.py

* refactor: replace list comprehension with set comprehension

Although there is nothing syntactically wrong with this code, it is hard to read and can be simplified to a set comprehension.
Using set comprehension is more performant since there is no need to create a transient list.

* refactor: remove length check in favour of truthiness of the object

Using the `len` function to check if a sequence is empty is not idiomatic and can be less performant than checking the truthiness of the object.  `len` doesn't know the context in which it is called, so if computing the length means traversing the entire sequence, it must; it doesn't know that the result is just being compared to 0.

* refactor: remove unnecessary use of comprehension

It is unnecessary to use a comprehension just to loop over the `iterable` and create a `list`/`set`/`dict` out of it. Python has a specialized set of tools for this task: the `list`/`set`/`dict` constructors, which are faster and more readable.

* refactor: change methods not using its bound instance to staticmethods

The method doesn't use its bound instance. Decorate this method with `@staticmethod` decorator, so that Python does not have to instantiate a bound method for every instance of this class thereby saving memory and computation. Read more about staticmethods [here](https://docs.python.org/3/library/functions.html#staticmethod).

* refactor: remove empty block of code

In most cases, an empty body of `for`, `while` or `if` implies some piece of code is missing.
Such empty block must be either filled or removed.

* refactor: `raise NotImplementedError` in empty functions

The function/method has been left empty here, without any comment or docstring. This can cause confusion later on why this was left empty.

* refactor: replace multiple `==` checks with `in`

To check if a variable is equal to one of many values, combine the values into a tuple and check if the variable is contained `in` it instead of checking for equality against each of the values.
This is faster, less verbose, and more readable.

* Update component_meta.py

* Update typing.py

* Update dice.py

* refactor: merge collapsible `if` statements

Nested `if` statements can be collapsed into a single `if` statement by separating their condition using `and` operator.
Merging collapsible `if` statements increases the code's readability.

* refactor: move append items to list definition

Instead of appending items to a list just after creation, it should be refactored to add those items into the list definition itself.

* Update __init__.py

* refactor: merge collapsible `with` statements

Nested `with` statements can be merged into a single compound `with` statement.

This issue is raised only when there are no other statements between the nested `with` statements.
Merging collapsible `with` statements can decrease cognitive complexity, improving code readability.

* refactor: remove unused imports

An object has been imported but is not used anywhere in the file.
It should either be used or the import should be removed.

* Delete .deepsource.toml

---------

Co-authored-by: deepsource-io[bot] <42547082+deepsource-io[bot]@users.noreply.github.com>
Co-authored-by: deepsource-autofix[bot] <62050782+deepsource-autofix[bot]@users.noreply.github.com>